requirements.txt
tensorflow-gpu
scikit-learn
matplotlib
sklearn